% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvalImp.R
\name{EvalImp}
\alias{EvalImp}
\title{Evaluate the imputed data sets and select the best data set}
\usage{
EvalImp(Originaldata, ImputedSets ,Imputed.mean, Imputed.sd)
}
\arguments{
\item{Originaldata}{data frame of original data containing the missing values}

\item{ImputedSets}{list of imputed data frames}

\item{Imputed.mean}{data frame of the means of the imputed data sets}

\item{Imputed.sd}{data frame of the standard deviations of the imputed data sets}
}
\value{
The best data frame which mean and standard deviation are close to the original data
}
\description{
The function is evaluate the imputed data sets based on the mean and standard deviation
}
\author{
Mohamed Soudy \email{Mohmedsoudy2009@gmail.com}
}
