% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt_admixture.R
\name{melt_admixture}
\alias{melt_admixture}
\title{Melt Admixture}
\usage{
melt_admixture(prefix, famdir, qdir)
}
\arguments{
\item{prefix}{Prefix for the filenames.}

\item{famdir}{Directory of the plink .FAM file.}

\item{qdir}{Directory location for the Q files.}
}
\value{
Processed dataframe.
}
\description{
This function reads specified files and processes them.
}
