% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERO.R
\name{MERO}
\alias{MERO}
\title{Perform Monte Carlo Expectation Maximization Random Forest Imputation}
\usage{
MERO(Data, ntree = 100, Nsets = 5)
}
\arguments{
\item{Data}{a data matrix with missing values. The columns correspond to the variables and the rows to the observations.}

\item{ntree}{number of trees to grow in each forest.}

\item{Nsets}{number of simulations/ data sets to be generated.}
}
\value{
A list containing data sets and imputed means, and imputed standard deviation.
}
\description{
The function is used to impute the missing data using Monte Carlo Expectation Maximization Random Forest Imputation
}
\author{
Mohamed Soudy \email{Mohmedsoudy2009@gmail.com}
}
