% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCorrelateMean.R
\name{PlotCorrelateMean}
\alias{PlotCorrelateMean}
\title{Plot the correlation in scatter plot between original mean and imputed mean}
\usage{
PlotCorrelateMean(OriginalMean, ImputedMean)
}
\arguments{
\item{OriginalMean}{means of the original data}

\item{ImputedMean}{means of the imputed data}
}
\value{
The scatter plot
}
\description{
The function is used to plot the correlation between the imputed mean and original mean
}
