% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jm_longx.R
\name{theme_jm_longx}
\alias{theme_jm_longx}
\title{Custom ggplot2 Theme Function}
\usage{
theme_jm_longx(font_size)
}
\arguments{
\item{font_size}{Numeric value specifying the font size to be used for text elements in the plot. The default value is 10.}
}
\value{
A ggplot2 theme object that can be added to ggplot2 plot objects.
}
\description{
This function applies a custom ggplot2 theme with specific aesthetic modifications. It modifies the base black-and-white theme by adjusting the text elements including axis text, plot title, legend text, and legend title, and sets their font family to "Myriad". Additionally, it allows the user to specify the font size for these elements. The function also positions the legend at the top of the plot.
}
